Use contains to check token relevance to a cid, not balances

+++ src/multi/Constants.sol
@@ -1,4 +1,4 @@
 // SPDX-License-Identifier: BUSL-1.1
 pragma solidity ^0.8.27;
 
-uint256 constant MAX_TOKENS = 16;
+uint8 constant MAX_TOKENS = 16;
